# optimize: Eliminate duplicate quality checks and optimize Cloud Build performance