--- a/.github/workflows/pr-checks.yml
+++ b/.github/workflows/pr-checks.yml
@@ -86,13 +86,13 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: TypeScript type checking
+      - name: TypeScript type checking (strict)
         run: npm run typecheck:quick
         env:
           NODE_ENV: development
 
-      - name: ESLint code analysis
-        run: npm run lint
+      - name: ESLint code analysis (strict)
+        run: npm run lint -- --max-warnings 0
 
       - name: Run property consistency checks
         run: npm run check:consistency
@@ -126,11 +126,28 @@ jobs:
       - name: Generate schema
         run: npm run generate-schema:force
 
-      - name: Test build process
+      - name: Test build process (production simulation)
         run: npm run build
         env:
           NODE_ENV: production
           SKIP_ENV_VALIDATION: true
+
+      - name: Verify build artifacts
+        run: |
+          echo "📦 Verifying Next.js build output..."
+          if [ -d ".next" ]; then
+            echo "✅ .next directory created"
+            if [ -f ".next/BUILD_ID" ]; then
+              echo "✅ Build ID file exists"
+            fi
+            if [ -d ".next/static" ]; then
+              echo "✅ Static assets generated"
+            fi
+            echo "📊 Build verification complete"
+          else
+            echo "❌ Build failed - .next directory not found"
+            exit 1
+          fi
 
   auth-integration-test:
     name: OAuth Integration Tests