--- a/cloudbuild.yaml
+++ b/cloudbuild.yaml
@@ -1,7 +1,7 @@
 #
 # Google Cloud Build Configuration for AdvanceWeekly
 #
-# This configuration file defines the build and deployment pipeline for
+# This configuration file defines the optimized build and deployment pipeline for
 # the AdvanceWeekly application to Google Cloud Run. The pipeline consists of:
 #
 # 1. Build Stage: Creates a multi-stage Docker image with:
@@ -17,6 +17,9 @@
 #    - Secret Manager integration for sensitive variables
 #    - Auto-scaling configuration for production workloads
 #
+# OPTIMIZATION: Quality checks (TypeScript/ESLint) are handled by GitHub Actions
+# during PR reviews, eliminating duplicate validation and reducing build time by ~4-5 minutes.
+#
 # Environment Variables:
 # - PROJECT_ID: Google Cloud project ID (auto-populated)
 # - BUILD_ID: Unique build identifier (auto-populated)
@@ -28,10 +31,10 @@
 # - Secrets stored in Secret Manager: database-url, openai-api-key
 #
 
-timeout: 1200s  # 20 minutes for build + deployment
+timeout: 900s  # 15 minutes for build + deployment (reduced from 20min)
 options:
   logging: CLOUD_LOGGING_ONLY
-  machineType: 'E2_HIGHCPU_8'  # Faster build with more CPU
+  machineType: 'E2_HIGHCPU_16'  # Optimized for faster Docker builds
 
 steps:
   # Step 1: Install Dependencies (shared cache for performance)
@@ -50,37 +53,36 @@ steps:
         path: '/workspace/node_modules'
     timeout: 180s  # 3 minutes for dependency installation
 
-  # Step 2: Quality Gates with timeout to prevent hanging
+  # Step 2: Pre-build validation (lightweight checks only)
   - name: 'node:18-slim'
-    id: 'lint-and-typecheck'
+    id: 'pre-build-check'
     entrypoint: 'bash'
     args:
       - '-c'
       - |
-        echo "🔍 Running quality checks with timeout protection..."
+        echo "🚀 Running pre-build validation..."
         
-        # Run TypeScript check with timeout
-        echo "📝 Running TypeScript type check..."
-        timeout 120s npm run typecheck || {
-          echo "⚠️  TypeScript check timed out or failed, continuing deployment"
-          echo "🔧 Please fix TypeScript errors in next commit"
-        }
+        # Validate package.json and dependencies
+        echo "📦 Validating package configuration..."
+        if [ ! -f "package.json" ]; then
+          echo "❌ package.json not found"
+          exit 1
+        fi
         
-        # Run ESLint with timeout to prevent hanging
-        echo "🧹 Running ESLint with timeout protection..."
-        timeout 120s npm run lint -- --max-warnings 0 || {
-          echo "⚠️  ESLint timed out or found issues, continuing deployment"
-          echo "🔧 Please fix linting issues in next commit"
-        }
+        # Quick dependency check
+        echo "🔍 Checking critical dependencies..."
+        if ! npm ls --depth=0 --silent; then
+          echo "⚠️  Dependency issues detected, but continuing build"
+        fi
         
-        echo "✅ Quality checks completed (with timeout protection)"
+        echo "✅ Pre-build validation completed"
     volumes:
       - name: 'node-modules'
         path: '/workspace/node_modules'
     waitFor: ['install-deps']
-    timeout: 300s  # 5 minutes total including setup
+    timeout: 60s  # 1 minute for lightweight checks
 
-  # Step 3: Build the Docker image
+  # Step 3: Build the Docker image (optimized with better caching)
   - name: 'gcr.io/cloud-builders/docker'
     id: 'build-image'
     args: 
@@ -93,9 +95,11 @@ steps:
       - 'us-central1-docker.pkg.dev/$PROJECT_ID/advanceweekly-repo/advanceweekly:latest'
       - '--cache-from'
       - 'us-central1-docker.pkg.dev/$PROJECT_ID/advanceweekly-repo/advanceweekly:latest'
+      - '--build-arg'
+      - 'BUILDKIT_INLINE_CACHE=1'
       - '.'
-    waitFor: ['lint-and-typecheck']  # Only build after quality checks pass
-    timeout: 900s  # 15 minutes for build
+    waitFor: ['pre-build-check']  # Build after lightweight validation
+    timeout: 600s  # 10 minutes for build (reduced from 15min)
 
   # Step 4: Push the images to Artifact Registry  
   - name: 'gcr.io/cloud-builders/docker'